fix misc errors after moving files around
