+++ src/eo_api/components/__init__.py
@@ -1,6 +1,13 @@
 """Reusable workflow components exposed as API capabilities."""
 
-from . import routes as routes
-from . import services as services
-
-__all__ = ["routes", "services"]
+from . import (
+    routes, 
+    schemas, 
+    services, 
+    data_registry, 
+    data_retrieval, 
+    dhis2_datavalueset, 
+    download, 
+    spatial_aggregation, 
+    temporal_aggregation, 
+)
