+++ src/eo_api/components/dhis2_datavalueset/schemas/fastapi.py
@@ -1,4 +1,9 @@
-from pydantic import BaseModel
+from typing import Any
+
+from pydantic import BaseModel, Field
+
+from .datavalueset import Dhis2DataValueSetConfig
+from ...schemas import PeriodType
 
 
 class BuildDataValueSetRunRequest(BaseModel):
