+++ src/eo_api/components/download/__init__.py
@@ -1,1 +1,1 @@
-from . import *
+from . import routes, services, schemas
