+++ src/eo_api/components/download/schemas/download.py
@@ -1,4 +1,4 @@
-from pydantic import BaseModel
+from pydantic import BaseModel, ConfigDict
 
 
 class _DownloadDatasetStepConfig(BaseModel):
