+++ src/eo_api/components/temporal_aggregation/__init__.py
@@ -1,1 +1,1 @@
-from . import *
+from . import routes, services, schemas
