+++ src/eo_api/components/temporal_aggregation/services/temporal.py
@@ -6,8 +6,9 @@ from typing import cast
 
 import xarray as xr
 
-from ...data_manager.services.utils import get_time_dim
-from ..schemas import AggregationMethod, PeriodType
+from ..schemas.temporal import AggregationMethod
+from ...schemas import PeriodType
+from ...download.services.utils import get_time_dim
 
 _PERIOD_TO_FREQ: dict[PeriodType, str] = {
     PeriodType.HOURLY: "1h",
@@ -40,26 +41,26 @@ def temporal_aggregation_component(
 
 
 # from workflows engine
-def _run_temporal_aggregation(
-    *,
-    runtime: WorkflowRuntime,
-    request: WorkflowExecuteRequest,
-    dataset: dict[str, Any],
-    context: dict[str, Any],
-    step_config: dict[str, Any],
-) -> dict[str, Any]:
-    target_period_type = PeriodType(
-        str(step_config.get("target_period_type", request.temporal_aggregation.target_period_type))
-    )
-    method = AggregationMethod(str(step_config.get("method", request.temporal_aggregation.method)))
-    temporal_ds = runtime.run(
-        "temporal_aggregation",
-        component_services.temporal_aggregation_component,
-        dataset=dataset,
-        start=request.start,
-        end=request.end,
-        bbox=_require_context(context, "bbox"),
-        target_period_type=target_period_type,
-        method=method,
-    )
-    return {"temporal_dataset": temporal_ds}
+# def _run_temporal_aggregation(
+#     *,
+#     runtime: WorkflowRuntime,
+#     request: WorkflowExecuteRequest,
+#     dataset: dict[str, Any],
+#     context: dict[str, Any],
+#     step_config: dict[str, Any],
+# ) -> dict[str, Any]:
+#     target_period_type = PeriodType(
+#         str(step_config.get("target_period_type", request.temporal_aggregation.target_period_type))
+#     )
+#     method = AggregationMethod(str(step_config.get("method", request.temporal_aggregation.method)))
+#     temporal_ds = runtime.run(
+#         "temporal_aggregation",
+#         component_services.temporal_aggregation_component,
+#         dataset=dataset,
+#         start=request.start,
+#         end=request.end,
+#         bbox=_require_context(context, "bbox"),
+#         target_period_type=target_period_type,
+#         method=method,
+#     )
+#     return {"temporal_dataset": temporal_ds}
